/trunk, cisst3DUserInterface: added camera calibration file in /share and use cmnPath to locate file in examples.   Use size_t whenever possible (warnings on MS 64 bits).
git-svn-id: https://svn.lcsr.jhu.edu/cisst/trunk@4562 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ cisst/cisst3DUserInterface/code/ui3BehaviorBase.cpp
@@ -210,15 +210,15 @@ void ui3BehaviorBase::SetState(const StateType & newState)
 }
 
 
-void ui3BehaviorBase::OnStreamSample(svlSample * CMN_UNUSED(sample), int CMN_UNUSED(streamindex))
+void ui3BehaviorBase::OnStreamSample(svlSample * CMN_UNUSED(sample), size_t CMN_UNUSED(streamindex))
 {
     // Default implementation does nothing
 }
 
 
-int ui3BehaviorBase::AddStream(svlStreamType type, const std::string & streamname)
+size_t ui3BehaviorBase::AddStream(svlStreamType type, const std::string & streamname)
 {
-    int streamindex = Streams.size();
+    size_t streamindex = Streams.size();
     ui3VideoInterfaceFilter* filter = new ui3VideoInterfaceFilter(type, streamindex, this);
     CMN_ASSERT(filter);
 
@@ -232,7 +232,7 @@ int ui3BehaviorBase::AddStream(svlStreamType type, const std::string & streamnam
 }
 
 
-unsigned int ui3BehaviorBase::GetStreamWidth(const int streamindex, unsigned int channel)
+unsigned int ui3BehaviorBase::GetStreamWidth(const size_t streamindex, unsigned int channel)
 {
     if (streamindex < 0 || streamindex >= static_cast<int>(Streams.size())) return 0;
     return Streams[streamindex]->GetWidth(channel);
